% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotRows}
\alias{plotRows}
\title{Plot the rows of a dataframe / tibble as a bar plot}
\usage{
plotRows(x)
}
\arguments{
\item{x}{A dataframe or tibble}
}
\value{
a ggplot2 object
}
\description{
Plot the rows of a dataframe / tibble as a bar plot
}
\examples{
plotRows(tibble::tibble(a = rnorm(3), b = rnorm(3)))
}
