% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{probabilityToOdds}
\alias{probabilityToOdds}
\title{Convert a probability to an odds}
\usage{
probabilityToOdds(probability)
}
\arguments{
\item{probability}{a probability}
}
\value{
an odds
}
\description{
Convert a probability to an odds
}
